Add _current_repodata_bioconda_hack requirements-only package
Similarly to _current_repodata_hack, this lists several requirements
for packages in the Bioconda channel to ensure the required conda-forge
packages appear in conda-forge's current_repodata.json:

Various packages from the Bioconda channel have pinnings on various
conda-forge packages. From time to time, these pinned versions are not
the latest versions of these packages in conda-forge, so conda-forge's
current_repodata.json does not contain them. This can lead to much older
versions of the Bioconda packages (which may have reduced dependencies
hence not require the conda-forge packages in question) being selected,
leading to user disappointment. This _current_repodata_bioconda_hack
package depends on the particular versions of these conda-forge packages
that are of interesting to Bioconda, thus ensuring that they appear in
conda-forge's current_repodata.json and avoiding this problem.